// src/interfaces/dadosFront.ts
interface dadosFront {

    id: string;
    name: string;
    email: string;
    banca: string;
    contest_name: string;
    registration_number: string;
    competition_examiner: string;
    question_number: string;
    discipline: string;
    topic: string;
    contest_answer_sheet: string;
    type: string;
    bank_answer_sheet: string;
    desired_answer: string;
    question_text: string;
}

export default dadosFront;